Move naturals and greatest_common_divisor into com package
